Add files via upload
It is a version of the source code that may be used in Python IDE. It has no user interface. 
Before use, please add your directories and criteria in the lines where PLACE_HOLDER text is printed. Additional comments for the user are written in CAPS. Additional comments about the process are written in lowercase. 
Level of the words that should be excluded are also should be specified before usage. Please see comments on lines 70-71. 

Tagger: Mecab.